api_kaddo/requirements.txt
black
celery==5.4.0
cs50==9.3.4
Django==5.0.6
django-cors-headers==4.4.0
django-oauth-toolkit==2.4.0
djangorestframework==3.15.2
djangorestframework-api-key==2.3.0
gunicorn==22.0.0
locust==2.31.2
openpyxl==3.1.5
pandas==2.2.2
python-decouple==3.8
psycopg2==2.9.9
pytest==8.3.2
redis==5.0.8
requests==2.32.3
Unidecode==1.3.8